Implement Localization for WebUI
Add localization support to the WebUI by implementing a language selection feature. This allows users to switch between different languages for the interface, enhancing accessibility and user experience.
This is a commit message for implementing localization in the WebUI of the project. The changes include:
- Added a new JavaScript file `i18n.js` to handle internationalization.
- Created language files for English, Spanish, and French in the `languages` directory.
- Updated the `index.html` and `custom.html` files to include language selection options.
- Modified the `index.js` file to load the appropriate language file based on user selection.
- Updated the `vite.config.js` file to include the new language files in the build process.

If you want to contribute to add your own language, please follow the structure of the existing language files and add your translations in the `languages` directory. Make sure to test the implementation thoroughly before submitting a pull request.
also the name of the xml files should be the same as the language code, for example:
- For English, the file should be named `en.xml`
- For Spanish, the file should be named `es.xml`