scripts: action: implement susfs4ksu userspace tool update script
This commit updates the susfs4ksu userspace tool to the latest version.
This is useful for users who updates their kernel to the latest version of SUSFS so that it won't be needed to reinstall the susfs4ksu module again.
It uses susfsd to check the SUSFS version in the kernel.