Rewritten Contributor and Translator generation.
The problem with the old system was that it didn't show when it's highway transition.
To solve this, use export function in the credits.js and import those functions and call it on the highway transition.
To solve also the characters not displaying, use the import of image paths same as the space.js and use indexes for the characters in the json file.

for reference, here's the indexes for the characters:
blue, 0
brown, 1
cyan, 2
orange, 3
yellow, 4
green, 5
lime, 6
pink, 7
purple, 8
red, 9
white, 10
black, 11