scripts: post-mount/boot-completed: Move custom sus_path execution to boot-completed.sh
To reduce the chances of issues with the post-mount script especially bootlooping or system instability caused by improper use of custom sus_path or a module that automatically adds the paths to the custom sus_path, the custom sus_path execution is moved to boot-completed stage.
But still it's highly not recommended to use custom sus_path for custom rom hiding, as it can cause issues with the system if not used properly.