scripts: improve SUSFS logging and stats Instead of using sleep function to get the SUSFS Log and stats in service.sh, execute the dmesg command in each boot stage (post-fs-data, post-mount, boot-completed) to get the SUSFS Log and stats.
The way it works is to execute the dmesg command in post-fs-data first as it's the first one to execute.
Then after dmesg command is executed from the post-fs-data, the last timestamp of the dmesg command is saved to a checkpoint file.
Then in post-mount, service, and boot-completed, the dmesg is executed again, but this time it will only show the logs after the last timestamp saved in the checkpoint file.

In this way, the SUSFS Log collection is more robust and does not depend fully on the sleep function (except in boot-completed).
This may reduce the chance of missing SUSFS logs in the boot process which will lead to use the fallback error in the WebUI.