scripts/customize/binaries: appropriately check susfs implementation by checking KernelSU version
Since KernelSU v2.0.0, the susfs implementation has changed to use supercalls starting from susfs v2.0.0.
This commit updates the customize.sh script to check the KernelSU version and select the appropriate susfs binary accordingly.
In this way, the correct susfs binary is used based on whether the KernelSU version is before or after v2.0.0.
Therefore, the local susfs binaries are split into two folders (pre-20000 and 20000) to accommodate this change.